修复容器每次启动时都要去生成license，却找不到生成的class类而产生的BUG

+++ start.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 
+activationfile="/ActivationCode.class"
 webdir="/u01/webimport/tomcat/"
 host="192.168.39.135"
 
 # 替换License
-oldCompany=`cat /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml | grep lien.company | awk -F " " '{print $2}'`
-oldContext=`cat /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml | grep lien.context | awk -F " " '{print $2}'`
-newCompany="wangjun"
-newContext=`java -cp / ActivationCode`
-sed -i "s|$oldCompany|$newCompany|g" /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml
-sed -i "s|$oldContext|$newContext|g" /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml
-
-# 删除java的二进制文件
-rm /ActivationCode.class
+if [ -f "$activationfile" ]; then
+  oldCompany=`cat /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml | grep lien.company | awk -F " " '{print $2}'`
+  oldContext=`cat /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml | grep lien.context | awk -F " " '{print $2}'`
+  newCompany="wangjun"
+  newContext=`java -cp / ActivationCode`
+  sed -i "s|$oldCompany|$newCompany|g" /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml
+  sed -i "s|$oldContext|$newContext|g" /u01/webimport/tomcat/webapps/ime-container/WEB-INF/classes/application.yaml
+  # 删除java的二进制文件
+  rm /ActivationCode.class
+fi
 
 # 替换数据库连接字符串
 sed -i "s|@127.0.0.1:1521:ora11g|@$host:1521:XE|g" `find /u01/webimport/tomcat/webapps/ -type f `
